pyproject.toml
[project]
name = "touchdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ankur Rustagi", email = "arustag@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.11.0",
    "ffmpeg-python>=0.2.0",
    "openai>=1.109.1",
    "yt-dlp>=2025.9.23",
]


